lib/ukvmem: Move vma_types.h under uk/vmem
This change moves the vma_types.h header out of the public uk/ header
namespace and under the appropriate lib prefix uk/vmem/.
